refactoring export of cryptoHash function
