+++ blockchain/block.js
@@ -1,6 +1,6 @@
-const hexToBinary = require('hex-to-binary')
+const hexToBinary = require('hex-to-binary');
 const { GENESIS_DATA, MINE_RATE } = require('../config');
-const cryptoHash = require('../util/crypto-hash');
+const { cryptoHash } = require('../util');
 
 class Block {
   constructor({ timestamp, lastHash, hash, data, nonce, difficulty }) {
@@ -30,7 +30,9 @@ class Block {
         timestamp,
       });
       hash = cryptoHash(timestamp, lastHash, data, nonce, difficulty);
-    } while (hexToBinary(hash).substring(0, difficulty) !== '0'.repeat(difficulty));
+    } while (
+      hexToBinary(hash).substring(0, difficulty) !== '0'.repeat(difficulty)
+    );
 
     return new this({
       timestamp,
@@ -45,7 +47,7 @@ class Block {
   static adjustDifficulty({ originalBlock, timestamp }) {
     const { difficulty } = originalBlock;
 
-    if (difficulty < 1) return 1
+    if (difficulty < 1) return 1;
     if (timestamp - originalBlock.timestamp > MINE_RATE) return difficulty - 1;
     return difficulty + 1;
   }
