+++ blockchain/block.test.js
@@ -1,7 +1,7 @@
-const hexToBinary = require('hex-to-binary')
+const hexToBinary = require('hex-to-binary');
 const Block = require('./block');
 const { GENESIS_DATA, MINE_RATE } = require('../config');
-const cryptoHash = require('../util/crypto-hash');
+const { cryptoHash } = require('../util');
 
 describe('Block', () => {
   const timestamp = 2000;
@@ -74,9 +74,9 @@ describe('Block', () => {
     });
 
     it('sets a `hash` that matches the difficulty criteria', () => {
-      expect(hexToBinary(minedBlock.hash).substring(0, minedBlock.difficulty)).toEqual(
-        '0'.repeat(minedBlock.difficulty)
-      );
+      expect(
+        hexToBinary(minedBlock.hash).substring(0, minedBlock.difficulty)
+      ).toEqual('0'.repeat(minedBlock.difficulty));
     });
 
     it('adjust the difficulty', () => {
@@ -110,7 +110,7 @@ describe('Block', () => {
 
     it('has a lower limit of 1', () => {
       block.difficulty = -1;
-      
+
       expect(Block.adjustDifficulty({ originalBlock: block })).toEqual(1);
     });
   });
