+++ util/index.js
@@ -1,12 +1,12 @@
 const EC = require('elliptic').ec;
-const cryptoHash = require('./crypto-hash')
+const cryptoHash = require('./crypto-hash');
 
 const ec = new EC('secp256k1');
 
 const verifySignature = ({ publicKey, data, signature }) => {
-  const keyFromPublic = ec.keyFromPublic(publicKey, 'hex')
+  const keyFromPublic = ec.keyFromPublic(publicKey, 'hex');
 
-  return keyFromPublic.verify(cryptoHash(data), signature)
+  return keyFromPublic.verify(cryptoHash(data), signature);
 };
 
-module.exports = { ec, verifySignature };
+module.exports = { ec, verifySignature, cryptoHash };
